create separate component for select header button
