+++ app/screens/home/search/results/header.tsx
@@ -2,76 +2,18 @@
 // See LICENSE.txt for license information.
 import React, {useCallback, useState} from 'react';
 import {useIntl} from 'react-intl';
-import {Text, View} from 'react-native';
-import Button from 'react-native-button';
+import {View} from 'react-native';
 
 import {useTheme} from '@context/theme';
 import {changeOpacity, makeStyleSheetFromTheme} from '@utils/theme';
+
+import SelectButton from './header_button';
 
 type Props = {
     onHeaderSelect: (val: string) => void;
     numberFiles: number;
     numberMessages: number;
 }
-
-type ButtonProps = {
-    onPress: () => void;
-    selected: boolean;
-    text: string;
-}
-
-const Header = ({onHeaderSelect, numberFiles, numberMessages}: Props) => {
-    const theme = useTheme();
-    const styles = getStyleFromTheme(theme);
-    const intl = useIntl();
-
-    const messagesText = intl.formatMessage({id: 'screen.search.header.messages', defaultMessage: 'Messages'});
-    const filesText = intl.formatMessage({id: 'screen.search.header.files', defaultMessage: 'Files'});
-
-    const [tab, setTab] = useState(0);
-
-    const handleMessagesPress = useCallback(() => {
-        onHeaderSelect('message-tab');
-        setTab(0);
-    }, [onHeaderSelect]);
-
-    const handleFilesPress = useCallback(() => {
-        onHeaderSelect('file-tab');
-        setTab(1);
-    }, [onHeaderSelect]);
-
-    const SelectButton = ({selected, onPress, text}: ButtonProps) => {
-        return (
-            <Button
-                containerStyle={[styles.button, selected && styles.selectedButton]}
-                onPress={onPress}
-            >
-                <Text style={[styles.text, selected && styles.selectedText]}>
-                    {text}
-                </Text >
-            </Button>
-        );
-    };
-
-    return (
-        <>
-            <View style={styles.container}>
-                <SelectButton
-                    selected={tab === 0}
-                    onPress={handleMessagesPress}
-                    text={`${messagesText} (${numberMessages})`}
-                />
-                <SelectButton
-                    selected={tab === 1}
-                    onPress={handleFilesPress}
-                    text={`${filesText} (${numberFiles})`}
-                />
-            </View>
-            <View style={styles.divider}/>
-        </>
-
-    );
-};
 
 const getStyleFromTheme = makeStyleSheetFromTheme((theme) => {
     return {
@@ -106,6 +48,46 @@ const getStyleFromTheme = makeStyleSheetFromTheme((theme) => {
         },
     };
 });
+
+const Header = ({onHeaderSelect, numberFiles, numberMessages}: Props) => {
+    const theme = useTheme();
+    const styles = getStyleFromTheme(theme);
+    const intl = useIntl();
+
+    const messagesText = intl.formatMessage({id: 'screen.search.header.messages', defaultMessage: 'Messages'});
+    const filesText = intl.formatMessage({id: 'screen.search.header.files', defaultMessage: 'Files'});
+
+    const [tab, setTab] = useState(0);
+
+    const handleMessagesPress = useCallback(() => {
+        onHeaderSelect('message-tab');
+        setTab(0);
+    }, [onHeaderSelect]);
+
+    const handleFilesPress = useCallback(() => {
+        onHeaderSelect('file-tab');
+        setTab(1);
+    }, [onHeaderSelect]);
+
+    return (
+        <>
+            <View style={styles.container}>
+                <SelectButton
+                    selected={tab === 0}
+                    onPress={handleMessagesPress}
+                    text={`${messagesText} (${numberMessages})`}
+                />
+                <SelectButton
+                    selected={tab === 1}
+                    onPress={handleFilesPress}
+                    text={`${filesText} (${numberFiles})`}
+                />
+            </View>
+            <View style={styles.divider}/>
+        </>
+
+    );
+};
 
 export default Header;
